Changed behavior  stop/error/warning handling.
The firmware now handles door open and chiller off events differently.
Door open used to  pause processing. Now it keeps processing but posts
a door open warning which can be shown in the GUI. The hard logic
system is now the only thing disabling the laser while the gantry keeps
moving.

Similarly a chiller off event does not invoke stop-mode. The gantry
keeps moving and a warning is posted to be shown by the GUI. This
allows testing of the mechanics without messing with the chiller
interlock that is crucial for preventing the operation of the laser
without cooling.